third_party/blink/web_tests/fast/events/window-onerror-isolatedworld-02-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE ERROR: Uncaught Error: Error in main world inline script.
CONSOLE ERROR: Uncaught Error: Error in main world load handler.
Test that window.onerror and "error" event listeners from isolated world are invoked for uncaught exceptions in scripts running in isolate worlds as well as for exceptions in the main world.Bug 8519.

isolated world window.onerror: Uncaught Error: Error in isolated world inline script. at window-onerror-isolatedworld-02.html, Line: 41, Column: 13
Error object present!
isolated world error event listener: Uncaught Error: Error in isolated world inline script. at window-onerror-isolatedworld-02.html:, Line: 41
Error object present!
isolated world window.onerror: Uncaught Error: Error in isolated world load handler. at window-onerror-isolatedworld-02.html, Line: 38, Column: 17
Error object present!
isolated world error event listener: Uncaught Error: Error in isolated world load handler. at window-onerror-isolatedworld-02.html:, Line: 38
Error object present!
isolated world window.onerror: Uncaught Error: Error in isolated world setTimeout callback. at window-onerror-isolatedworld-02.html, Line: 36, Column: 21
Error object present!
